==> notebooks/note/model/3_model_verif_label_shuffle.ipynb <==
import pandas as pd
import joblib
from xgboost import XGBClassifier
from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score


pos

df_pos2x = pd.read_csv('../../data/basedata/df_pos_MCH.csv')
columns_to_drop = ['Metabolitename', 'Ontology', 'dataset', 'AlignmentID', 'AverageMz']
X = df_pos2x.drop(columns=columns_to_drop).astype(float).values
y = df_pos2x['Ontology']
X_train, X_test_pos, y_train, y_test_pos = train_test_split(X, y, test_size=0.2, random_state=42, stratify=y)

loaded_model_pos = joblib.load('../../data/model_comp/pos_pred_result/best_xgb_model_random_pos.joblib')
label_encoder_pos = joblib.load('../../data/model_comp/pos_pred_result/label_encoder_pos_random.pkl')
y_test_pos = label_encoder_pos.transform(y_test_pos)

y_pred_pos = loaded_model_pos.predict(X_test_pos)
print('XGBoost Accuracy -positive ion mode:', accuracy_score(y_test_pos, y_pred_pos))

label shuffle

import numpy as np
y_random = np.random.choice(df_pos2x['Ontology'].unique(), size=len(df_pos2x), replace=True)
X_train, X_test, y_train_pos_random, y_test_pos_random = train_test_split(X, y_random, test_size=0.2, random_state=42, stratify=y_random)
y_test_pos_random = label_encoder_pos.transform(y_test_pos_random)
y_train_pos_random = label_encoder_pos.transform(y_train_pos_random)

params_pos = loaded_model_pos.get_xgb_params()
model_pos_ramdon = XGBClassifier(**params_pos)
model_pos_ramdon.fit(X_train, y_train_pos_random)

y_pred_pos_ranodm = model_pos_ramdon.predict(X_test)
print('XGBoost Accuracy with label shuffling -positive ion mode:', accuracy_score(y_test_pos_random, y_pred_pos_ranodm))

neg

df_neg2x = pd.read_csv('../../data/basedata/df_neg_MCH.csv')
columns_to_drop = ['Metabolitename', 'Ontology', 'dataset', 'AlignmentID', 'AverageMz']
X = df_neg2x.drop(columns=columns_to_drop).astype(float).values
y = df_neg2x['Ontology']
X_train, X_test_neg, y_train, y_test_neg = train_test_split(X, y, test_size=0.2, random_state=42, stratify=y)

loaded_model_neg = joblib.load('../../data/model_comp/neg_pred_result/best_xgb_model_random_neg.joblib')
label_encoder_neg = joblib.load('../../data/model_comp/neg_pred_result/label_encoder_neg_random.pkl')
y_test_neg = label_encoder_neg.transform(y_test_neg)

y_pred_neg = loaded_model_neg.predict(X_test_neg)
print('XGBoost Accuracy -negative ion mode:', accuracy_score(y_test_neg, y_pred_neg))

import numpy as np
y_random = np.random.choice(df_neg2x['Ontology'].unique(), size=len(df_neg2x), replace=True)
X_train, X_test, y_train_neg_random, y_test_neg_random = train_test_split(X, y_random, test_size=0.2, random_state=42, stratify=y_random)
y_test_neg_random = label_encoder_neg.transform(y_test_neg_random)
y_train_neg_random = label_encoder_neg.transform(y_train_neg_random)

params_neg = loaded_model_neg.get_xgb_params()
model_neg_ramdon = XGBClassifier(**params_neg)
model_neg_ramdon.fit(X_train, y_train_neg_random)

y_pred_neg_ranodm = model_neg_ramdon.predict(X_test)
print('XGBoost Accuracy with label shuffling -negative ion mode:', accuracy_score(y_test_neg_random, y_pred_neg_ranodm))
